% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcol.R
\name{addcol}
\alias{addcol}
\title{Add a new column to a data frame and fill with a value.}
\usage{
addcol(data, col, value)
}
\arguments{
\item{data}{A data frame}

\item{col}{The column name to add (character, quoted)}

\item{value}{The value to fill the column with, quoted for character vector.}
}
\value{
a data frame
}
\description{
Use to add an identifying column to a data frame. For example,
  add a column called species and fill the column with the name of the
  species.
}
\examples{
\dontrun{df <- addcol(data, "species", "myrtus communis") }
}

