% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_df.R
\name{character_df}
\alias{character_df}
\title{Convert data.frame columns to character}
\usage{
character_df(data)
}
\arguments{
\item{data}{A data.frame with factor or list columns.}
}
\value{
a data.frame
}
\description{
Useful where columns in a data.frame include factors or lists that you don't want to deal with at the time (e.g. exploring an API return). Helpful if you want to use readr::write_csv() but get blocked from doing so by the existence of lists in the data.frame.
}
\details{
From an SO answer by Hadley Wickham \url{http://stackoverflow.com/questions/2851015/convert-data-frame-columns-from-factors-to-characters}
}
\examples{
\dontrun{character_df(data)}
}

