% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref_authors.R
\name{crossref_authors}
\alias{crossref_authors}
\title{Create data.frame with authors from crossref}
\usage{
crossref_authors(data, source, .id = id)
}
\arguments{
\item{data}{x$author. The author list in a data.frame produced by
\code{get_crossref}}

\item{source}{The source crossref data.frame name (unquoted) produced as
output by \code{get_crossref}.}

\item{.id}{This must be the DOI field from the source crossref data.frame as
x$DOI (unquoted).}
}
\value{
a data.frame
}
\description{
This function builds on rcrossref \code{cr_works}. It takes a
  data.frame and extracts the `author` list column into a data.frame with one
  author per line and a new `full_name` column (family, given).
}
\details{
A data.frame of results from crossref (see \code{get_crossref} )
  containing an author column in list format with author family and given
  names. This function extracts authors one per row per DOI. The DOI from the
  source table serves as the index. The extracted author data is joined with
  the data in the source table using dplyr \code{leftjoin} to create a new
  data.frame with an author name per row per DOI. A new `full_name` column is
  added as `family, given` using tidyr \code{unite}. The data can then be
  used for author name cleaning, analysis and visualization. To split other
  data columns for further processing use tidyr \code{separate_rows}. Warning
  At present this version only addresses crossref data where a DOI column is
  present and does not address other identifiers (ISBN, ISSN). A DOI column
  must be present for the function to work.
}
\examples{
\dontrun{pizza_authors <- crossref_authors(pizza_crossref_df$author, pizza_crossref_df, .id = pizza_crossref_df$DOI)}
}
\references{
\url{https://github.com/ropensci/rcrossref}
}
