% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref_count.R, R/crossref_count_test.R
\name{crossref_count}
\alias{crossref_count}
\title{Count results for query using rcrossref}
\usage{
crossref_count(query)

crossref_count(query)
}
\arguments{
\item{query}{A search term}

\item{query}{}
}
\value{
numeric

numeric
}
\description{
A simple wrapper for rcrossref::cr_works() that extracts the total number of results from the meta data for a query. Useful for dividing up the calls within the 1000 limit using the offset argument in cr_works. e.g Call 1. cr_works(query, limit = 1000), Call 2. cr_works(query, limit = 1000, offset = 1001) where Call 2 retrievs the next 1000 results and so on.

A simple wrapper for rcrossref::cr_works() that extracts the total number of results from the meta data for a query. Useful for dividing up the calls within the 1000 limit using the offset argument in cr_works. e.g Call 1. cr_works(query, limit = 1000), Call 2. cr_works(query, limit = 1000, offset = 1001) where Call 2 retrievs the next 1000 results and so on.
}
\examples{
\dontrun{pizza <- crossref_count(query = "pizza")}
\dontrun{pizza <- crossref_count(query = "pizza")}
}

