% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref_rename.R
\name{crossref_rename}
\alias{crossref_rename}
\title{Rename crossref fields}
\usage{
crossref_rename(x)
}
\arguments{
\item{x}{A data.frame with columns to rename}
}
\value{
A data.frame
}
\description{
Start of a function to rename fields generated from rcrossref \code{cr_works} with underscores. Does not yet address complete field names.
}
\examples{
\dontrun{crossref_rename}
}
\references{
\url{https://github.com/ropensci/rcrossref}
}

