% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derwent_extract.R
\name{derwent_extract}
\alias{derwent_extract}
\title{Extract sections of Derwent patent full texts to columns}
\usage{
derwent_extract(x, col, section, remove = TRUE)
}
\arguments{
\item{x}{a data.frame}

\item{col}{A column in x}

\item{section}{either number, title, abstract, description, or claims or "tac" or "al;l"}

\item{remove}{Whether to remove the input column. Default is true.}
}
\value{
A data.frame with new columns
}
\description{
Given a data frame containing the full texts of patent texts, this function provides options to split the texts into number, title, abstract, description or claims. Options are also provided to just break out the title abstract or claims ("tac") or all fields.
}
\examples{
\dontrun{derwent_extract(five_texts, "tac")
derwent_extract(five_texts, "all")}
}
