% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derwent_format.R
\name{derwent_format}
\alias{derwent_format}
\title{Derwent Innovation Text file name matcher}
\usage{
derwent_format(x, col)
}
\arguments{
\item{x}{A data frame}

\item{col}{A column containing long form publication numbers}
}
\value{
data.frame
}
\description{
Downloads of text files in Derwent Innovation take a long form including the date and extension txt (e.g. EP1224299A220020724.txt). There is no consistent separator between the publication number and file names are not a consistent length. This prevents joining with table format downloads. This function reformats the numbers to solve this problem.
}
\details{
Derwent Innovation downloads of patents texts in txt format use patent numbers in long form. These numbers are not a uniform length. However, the YYYYMMDD.txt chunk of the string is a uniform 12 characters. The function calculates the length of each string and deducts the last 12 characters to arrive at the publication number for matching with table data.
}
\examples{
\dontrun{df <- derwent_format(derwent_text_format, "doc_id")}
}
