% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicates.R
\name{detect_duplicates}
\alias{detect_duplicates}
\title{Detect all duplicates in a column or vector}
\usage{
detect_duplicates(x)
}
\arguments{
\item{x}{a data frame, column or vector}
}
\value{
data.frame
}
