% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_five_texts.R
\docType{data}
\name{five_texts}
\alias{five_texts}
\title{five_texts}
\format{An object of class \code{readtext} (inherits from \code{data.frame}) with 5 rows and 2 columns.}
\usage{
five_texts
}
\description{
A data frame with two columns, doc_id and text for five patent full texts downloaded from derwent innovation.
}
\keyword{datasets}
