% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatto_derwent.R
\name{format_toderwent}
\alias{format_toderwent}
\title{convert publication number to derwent}
\usage{
format_toderwent(x, col)
}
\arguments{
\item{x}{a data.frame}

\item{col}{A column containing publication numbers to be transformed.}
}
\value{
a data frame containing the original publication numbers and a column in derwent format. Could be updated to tidyeval
}
\description{
Work in progress function to convert publication numbers from PATSTAT to Derwent Innovation format. This involves identifying EP publications with a zero at the start of a number and removing them. For US numbers those numbers that are 14 charcaters are converted to 15 characters through the addition of a padding zero after CCYYYY.
}
\details{
Be careful when using this function as it is not thoroughly tested. Also be careful that for existing mixed datasets conversion may duplicate an existing derwent innovation number.
}
\examples{
\dontrun{format_toderwent(problem_numbers)}
}
