% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonames_single.R
\name{geonames_single}
\alias{geonames_single}
\title{Read in a file downloaded from geonames and regularise the column names}
\usage{
geonames_single(path)
}
\arguments{
\item{path}{}
}
\value{
data.frame
}
\description{
Assumes a file has been downloaded from the geonames dump, imports with the correct column type and assigns the correct name.
}
