% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonames_prepare.R
\name{geonames_table}
\alias{geonames_table}
\title{Extract geonames export dump file names}
\usage{
geonames_table(...)
}
\value{
data.frame
}
\description{
Fetch the geonames daily dump page and parse it to a data frame.
}
\details{
The function fetches the geoname daily dump page consisting of a table with "Name", "Last Modified", "Size" and "Description". The Name field is parsed to create a file_name and file_type column. The files are a mix of individual country files (e.g. AD.zip) and other files (e.g. allCountries.zip, admin2Codes.txt etc.). To facilitate country matches the file name is parsed into iso2c for two letter country codes and other
}
\examples{
{geonames_prepare()}
}
