% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gephi_prepare.R
\name{gephi_prepare}
\alias{gephi_prepare}
\title{Prepare data for network visualization in gephi}
\usage{
gephi_prepare(x, col, count_col)
}
\arguments{
\item{x}{a .csv format file containing the data to visualize}

\item{col}{The column you would like to network}

\item{count_col}{The column used as the basis for calculating shared records as edges (for example dois, isi numbers, publication_numbers, priority_numbers, family_numbers).}
}
\value{
A data.frame
}
\description{
This function imports a .csv file, selects the desired columns, trims white space and adds a weight field based on the number of rows. Use the output to create a nodes file (gephi_nodes()) and edges file (gephi_edges()) for import into gephi.
}
