% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crossref.R
\name{get_crossref}
\alias{get_crossref}
\title{get_cross ref dois}
\usage{
get_crossref(dois)
}
\arguments{
\item{dois}{A vector of dois}
}
\value{
A data frame
}
\description{
A simple wrapper for rcrossref \code{cr_works} that returns a data.frame dropping other data. Note that this wrapper does not include the additional arguments available for \code{cr_works}.
}
\examples{
\dontrun{get_crossref(dois)}
}
\references{
\url{https://github.com/ropensci/rcrossref}
}
