% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_links.R
\name{get_google_links}
\alias{get_google_links}
\title{Retrieve links from a google search}
\usage{
get_google_links(url)
}
\arguments{
\item{url}{A search url created with google_url()}
}
\value{
A list of urls requiring further parsing (at the moment)
}
\description{
Use a url created with google_url() to retrieve a list of links
  for the research results. This function needs combining into one and to
  handle parsing of the urls returned.
}
\details{
The URLS returned need to be parsed at the beginning after `=`` and
  then before `&sa`.

from this SO question.
  \url{http://stackoverflow.com/questions/32889136/how-to-get-google-search-results}
}
\examples{
get_google_links(url) \\not run
}

