% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrows_listdf.R
\name{getrows_listdf}
\alias{getrows_listdf}
\title{Get n rows from a list of data.frames}
\usage{
getrows_listdf(x, rows = NULL)
}
\arguments{
\item{x}{A list of data.frames such as one created using `split(.,
asean_categories [, 1])`}

\item{rows}{The number of rows to return per data.frame (e.g. 1:5)}
}
\value{
a list of data frames with the corresponding answer
}
