% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_url.R
\name{google_url}
\alias{google_url}
\title{Create a Google Search Url}
\usage{
google_url(search_term, domain = ".co.uk", quotes = TRUE)
}
\arguments{
\item{search_term}{A search query}

\item{domain}{default is .co.uk. Change to your local service as needed}

\item{quotes}{whether the search should be quoted or not (default is TRU)}
}
\value{
a URL
}
\description{
enter a search term to generate a url to search google. Use with goog
}
\details{
from this SO question.
  \url{http://stackoverflow.com/questions/32889136/how-to-get-google-search-results}
}
\examples{
\dontrun{google_url(pizza)}
}

