% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inparallel.R
\name{inparallel}
\alias{inparallel}
\title{Run code in parallel}
\usage{
inparallel(cores = NULL, df, packages = "NULL", value = "NULL",
  expr = "NULL")
}
\arguments{
\item{cores}{Number of cores to use in the cluster}

\item{df}{A data frame}

\item{packages}{A vector of one or more packages}

\item{value}{A vector of values}
}
\description{
Use \code{multidplyr} to set up a local cluster to run code in parallel. Packages automatically transferred are the tidyverse and tidytext
}
