% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_numbers.R
\name{lens_numbers}
\alias{lens_numbers}
\title{Remove spaces and extra characters in patent numbers from the Lens.}
\usage{
lens_numbers(df, col)
}
\arguments{
\item{df}{A data.frame containing patent numbers}

\item{col}{the character column (quoted) containing the publication numbers}
}
\value{
a data.frame
}
\description{
Removes spaces after country codes and before kind codes along
  with "/" in patent numbers in data from the Lens. Useful for harmonising
  numbers for work with espacenet, Open Patent Services and other databases.
}
\details{
Note that the Lens is planning a facility that will allow sets of
  patent numbers to be uploaded. It can be assumed that the same format will
  be required as for data downloaded from the Lens. It will thefore often be
  prudent to create a new column with the edited number as in example 2
  below.
}
\examples{
\dontrun{df$new_col <- lens_numbers(df, "publication_number")}
}

