% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref_links.R
\name{links}
\alias{links}
\title{Extract crossref link data into a data.frame}
\usage{
links(col, source)
}
\arguments{
\item{col}{the list column of a data.frame containing cross ref link}

\item{source}{The source data.frame}
}
\value{
A data.frame
}
\description{
Takes the return from rcrossref \code{cr_works} and extracts the data in the link field and joins it to the source data.frame.
}
\details{
A DOI (quoted) column must be present in the source data.frame.
}
\examples{
\dontrun{links(col, source)}
}
\references{
\url{https://github.com/ropensci/rcrossref}
}

