% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_derwentnumbers.R
\name{ops_thomson_redraft}
\alias{ops_thomson_redraft}
\title{reformat Thomson Innovation (Derwent Innovation) patent numbers}
\usage{
ops_thomson_redraft(x, col)
}
\arguments{
\item{x}{A data.frame containing patent numbers in Thomson Innovation/Derwent Innovation Format}
}
\value{
A tibble with formatted numbers
}
\description{
reformats Thomson Innovation (Clarivate Analtics Derwent Innovation) patent numbers for calls to the OPS service by removing extra zeros and correcting for the insertion of years on PCT numbers. See details.
}
\details{
Thomson Innovation patent numbers involve the insertion of extra zeros to make US patent numbers nchar 15 and will not be found by OPS. In the case of Patent Cooperation Treaty (WO) documents, the year is inserted for records prior to the 1st of January 2004. Thus the form WO98xxx becomes WO1998xxx and will not be found in a call to OPS. In a small number of cases (so far) the Thomson Innovation number deleted a leading zero from EP numbers e.g. EP0xxx becomes EPxx and will not be found. At present the function does not address this issue and results from a call to EP numbers with OPS should be checked against the input numbers.
}
\examples{
\dontrun{ops_thomson(inpadoc_first_family_members, col = "inpadoc_first_family_number")}
}
