% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_df.R
\name{orcid_df}
\alias{orcid_df}
\title{Extract data.frames from orcid_doi searches}
\usage{
orcid_df(dois = NULL, start = NULL, rows = NULL, fuzzy = FALSE)
}
\arguments{
\item{dois}{A vector of dois for use in orcid_doi.}

\item{start}{(integer) Result number to start on. Keep in mind that pages
start at 0. (rorcid documentation)}

\item{rows}{(integer) Numer of results to return.}

\item{fuzzy}{fuzzy    (logical) Use fuzzy matching on input DOIs. Defaults to
FALSE. If FALSE, we stick "digital-object-ids" before the DOI so that the
search sent to ORCID is for that exact DOI. If TRUE, we use some regex to
find the DOI.}
}
\value{
A data frame
}
\description{
This wraps the orcid_doi() function from rorcid and parses the
  results to return a data frame. Note that it drops results that are not
  found.
}
\details{
A convenience wrapper for the rorcid_doi function from rorcid. See
  the rorcid package \url{https://github.com/ropensci/rorcid} (from which the
  documentation is taken). The advantage of the function is simply that it
  returns a data.frame.
}
\examples{
\dontrun{orcid_df(dois)}
}
