% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_process.R
\name{orcid_process}
\alias{orcid_process}
\title{Process rorcid results into a data frame}
\usage{
orcid_process(x)
}
\arguments{
\item{x}{A list of results returned from rorcid orcid_doi().}
}
\value{
a data.frame
}
\description{
Turn raw results from orcid into a data frame
}
\details{
Note dependency on oldhammisc::orcid_rename
}
