% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_orcid_rename_output.R
\docType{data}
\name{orcid_rename_output}
\alias{orcid_rename_output}
\title{orcid_rename_output}
\format{An object of class \code{data.frame} with 90 rows and 45 columns.}
\usage{
orcid_rename_output
}
\description{
orcid_sample with renamed columns using orcid_rename.
}
\keyword{datasets}
