% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_orcid_sample.R
\docType{data}
\name{orcid_sample}
\alias{orcid_sample}
\title{orcid_sample}
\format{An object of class \code{data.frame} with 90 rows and 45 columns.}
\usage{
orcid_sample
}
\description{
A sample return of data from orcid (use dplyr::bind_rows()) containing 90 rows and 45 columns with raw names.
}
\keyword{datasets}

