% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pizza_crossref.R
\docType{data}
\name{pizza_crossref_df}
\alias{pizza_crossref_df}
\title{pizza_crossref_df}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 32 columns.}
\usage{
pizza_crossref_df
}
\description{
A data frame with 20 rows and 32 columns from a search of crossref for teh word pizza with rcrossref.
}
\keyword{datasets}

