% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plos_records.R
\name{plos_records}
\alias{plos_records}
\title{Retrieve count of results for a query with PLOS}
\usage{
plos_records(query)
}
\arguments{
\item{query}{A search term or vector of search terms. For multiple terms use double quotes (see examples).}
}
\value{
prints the maximum value of results
}
\description{
Use this function to work out how many results a query of PLOS using rplos will return. Useful for deciding on data to download.
}
\examples{
\dontrun{drones <- plos_records(drones)}
\dontrun{synbio <- c('"synthetic biology"', '"synthetic genomics"', '"synthetic genome"', '"synthetic genomes"')
synbio <- plos_records(synbio)}
}
