% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_problem_numbers.R
\docType{data}
\name{problem_numbers}
\alias{problem_numbers}
\title{problem_numbers}
\format{data}
\usage{
problem_numbers
}
\description{
A data frame with a column of problem_numbers from ABSPAT in PATSTAT format for conversion into Derwent Innovation format using \code{reformat_derwentnumbers}.
}
\keyword{datasets}
