% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_isi.R
\name{process_isi}
\alias{process_isi}
\title{(depracated - use tidyr::separate_rows). Select field in dataframe containing ISI and/or DOI, separate, gather and
trim}
\usage{
process_isi(data, x = "", sep = "")
}
\arguments{
\item{data}{A data frame containing the target concatenated field and isi or
doi}

\item{x}{A concatenated character column (authors, organisations etc)}

\item{sep}{a separator (normally `;`)}
}
\value{
a data frame containing the target (whitespace trimmed), isi and doi
  fields
}
\description{
For working with scientific literature containing concatenated
  author, country, organisation or other fields. The data.frame must contain
  isi or doi numbers or something similar for the function to work.
}
\details{
Prints the number of separated columns for information part way through the
  function.
}
\examples{
\dontrun{process_isi(df, x = "authors", sep = ";")}
}
