% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxl_online.R
\name{readxl_online}
\alias{readxl_online}
\title{Read an Excel file from a URL}
\usage{
readxl_online(url, type = NULL, ...)
}
\arguments{
\item{url}{A url containing an excel .xls or xlsx or .zip file extension}

\item{type}{.xls or .xlsx or other excel type (not tested)}

\item{...}{Other arguments to readxl::read_excel}
}
\value{
data.frame
}
\description{
Read an online excel file where the URL contains the file extension or a zip file. The type argument attempts to address cases where the file extension is known but is not found in the url itself.
}
\examples{
\dontrun{ dfxls <- readxl_online("https://github.com/wipo-analytics/opensource-patent-analytics/blob/master/2_datasets/pizza_all_24294/patentscope_pizza_1940_2005_9659.xls?raw=true")

dfxlsx <- readxl_online("https://github.com/wipo-analytics/opensource-patent-analytics/blob/master/2_datasets/ewaste/ewaste.xlsx?raw=true")

dfzip <- readxl_online("https://github.com/poldham/opensource-patent-analytics/blob/master/2_datasets/taxonomy_final.zip?raw=true")
}
}
