% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_brackets.R
\name{remove_brackets}
\alias{remove_brackets}
\title{Remove Text in Brackets}
\usage{
remove_brackets(x)
}
\arguments{
\item{x}{A vector or column in a data frame}
}
\value{
x A vector or data frame column
}
\description{
Uses string R to replace text in brackets
}
\examples{
\dontrun{stringr::str_replace(doubtful$scientificname, " \\\\(.*\\\\)", "")}
}
