% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sentences_country.R
\docType{data}
\name{sentences_country}
\alias{sentences_country}
\title{sentences_country}
\format{An object of class \code{data.frame} with 868 rows and 13 columns.}
\usage{
sentences_country
}
\description{
A data frame with 898 sentences from patent documents containing an asean country name from the ASEAN marine dataset.
}
\keyword{datasets}
