% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_count.R
\name{sep_count}
\alias{sep_count}
\title{Count Separators for Separate_Gather}
\usage{
sep_count(data, col = "", sep = "[^[:alnum:]]+")
}
\arguments{
\item{data}{A data frame}

\item{col}{column containing separators to count}

\item{sep}{separator}
}
\value{
numeric count of separators
}
\description{
Count the number of separators in a column as the basis for the calculation of the number of columns to separate a column into using separate() in tidyr. Provides input to separate_gather
}
\details{
Calculates the maximum number of separators in a column (typically `;`). When dividing values by a separator note that the final entry will not be marked by a separator. To address this the function uses +1. Caution should be exercised when using `,` as a separator when using person or organisation names containing this punctuation to separate first and second names etc.
}
\examples{
\dontrun{sep_count(data, col = "authors", sep = ";")}
}

