% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_gather.R
\name{separate_gather}
\alias{separate_gather}
\title{Separate column values then gather them back in with tidyr}
\usage{
separate_gather(data = x, col = "", into = NULL, sep = "[^[:alnum:]]+",
  key = NULL, new_col = "", from = NULL, na.rm = TRUE)
}
\arguments{
\item{data}{x = A data frame}

\item{col}{A character column to separate}

\item{into}{The number of columns to separate into}

\item{sep}{The separator}

\item{key}{used by gather(). Only specify if required.}

\item{new_col}{Name of the character column to gather the data into.}

\item{from}{numeric position of columns to gather into new-col}

\item{na.rm}{Logical. Should NA values be removed. Default = TRUE}
}
\value{
A dataframe
}
\description{
Combines the common task of separating multiple values in a single column into individual columns based on a separator and then gathering them back in again.
}

