% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_splitdf.R
\docType{data}
\name{splitdf}
\alias{splitdf}
\title{splitdf}
\format{An object of class \code{list} of length 4266.}
\usage{
splitdf
}
\description{
A list of 4266 patent documents for use in testing earliest_priority.
}
\keyword{datasets}
