% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_splitwo.R
\docType{data}
\name{splitwo}
\alias{splitwo}
\title{splitwo}
\format{An object of class \code{list} of length 1643.}
\usage{
splitwo
}
\description{
A list of 1643 patent documents for use in testing earliest_priority.
}
\keyword{datasets}
