% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_to_col2.R
\name{term_to_col2}
\alias{term_to_col2}
\title{Search a column to identify terms and add column}
\usage{
term_to_col2(df, col, expr)
}
\arguments{
\item{df}{A data frame with the target column}

\item{col}{The column to be searched}

\item{expr}{A search term or vector of search terms}
}
\description{
Takes a vector of search terms and searches the column of a data.frame. Returns named columns to the data frame with logical values.
}
\examples{
\dontrun{a <- term_to_col2(sentences_country, "word", expr = c("ocean", "marine"))}
}
