% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_inventor_match.R
\name{wos_inventor_match}
\alias{wos_inventor_match}
\title{Match WOS author names to USPTO inventors}
\usage{
wos_inventor_match(x)
}
\arguments{
\item{x}{A vector containing raw author names.}
}
\value{
data frame
}
\description{
This draft function takes a data frame of raw unprocessed names from Web of Science and processes them to match with the USPTO raw inventors patent table as a basis for matching to cleaned inventor names.
}
