% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_query.R
\name{wos_query}
\alias{wos_query}
\title{Search Web of Science}
\usage{
wos_query(topic = NULL, address = NULL, wos_category = NULL)
}
\arguments{
\item{topic}{A topic such as a country name}

\item{address}{A country name}

\item{wos_category}{A web of science category or vector or categories selected from data/wos_categories}
}
\value{
data.frame
}
\description{
This function allows for a search of Web of Science on single search terms using the topic or the address field and one or more Web of Science Subject Categories.
}
