% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_gather.R
\name{separate_gather}
\alias{separate_gather}
\title{Separate a concatenated column on a separator and gather into a new column}
\usage{
separate_gather(data = x, col = "", sep = "[^[:alnum:]]+", new_col = "",
  na.rm = TRUE)
}
\arguments{
\item{data}{x = A data frame}

\item{col}{A concatenated character column to separate using sep (quoted)}

\item{sep}{The separator e.g. `;`}

\item{new_col}{Name of the character column to gather the data into (quoted).}

\item{na.rm}{Logical. Should NA values be removed. Default = TRUE}
}
\value{
A dataframe with separated observations in new rows and a named value column.
}
\description{
Combines the common task of separating multiple values in a single column into individual columns based on a separator and then gathering them back in again.
}
\details{
The function uses separate_ and gather_ from tidyr. A function sep_count counts up the number of separators and +1 to calculate the number of columns to be used by separate. These are then gathered back in. The default for na.rm is set to TRUE.
}
\examples{
\dontrun{separate_gather(data = data, col = "authors", sep = ";", new_col = "authors_split", na.rm = TRUE)}
}

